+++ nginx/conf.d/default.conf
@@ -45,7 +45,7 @@ server {
     gzip_min_length 256;
 
     add_header X-Content-Type-Options "nosniff" always;
-    add_header X-Frame-Options "DENY" always;
+    add_header X-Frame-Options "SAMEORIGIN" always;
     add_header X-XSS-Protection "1; mode=block" always;
     add_header Referrer-Policy "strict-origin-when-cross-origin" always;
     add_header Permissions-Policy "camera=(), microphone=(), geolocation=()" always;
@@ -115,6 +115,10 @@ server {
         try_files $uri =404;
         expires 7d;
         add_header Cache-Control "public, no-transform";
+    }
+
+    location = /feedback {
+        try_files /feedback/feedback.html =404;
     }
 
     location / {
