+++ src/feedback/feedback.html
@@ -1,0 +1,24 @@
+<!DOCTYPE html>
+<html>
+<head>
+    <meta charset='utf-8'>
+    <title>Обратная связь</title>
+    <link rel="stylesheet" href="/app/index.css?v=2">
+    <style>
+        html, body {
+            margin: 0;
+            min-height: 100vh;
+            background: rgba(0, 0, 0, 0.5);
+        }
+        body {
+            display: flex;
+            align-items: center;
+            justify-content: center;
+        }
+    </style>
+</head>
+<body>
+<div id="feedback-root"></div>
+<script type="module" src="/feedback/index.js"></script>
+</body>
+</html>
