+++ src/feedback/index.js
@@ -1,0 +1,4 @@
+import { FeedbackPage } from './FeedbackPage.js';
+
+const page = new FeedbackPage(document.getElementById('feedback-root'));
+page.init();
