+++ src/widgets/feedback-modal/FeedbackModal.css
@@ -1,19 +1,15 @@
-.feedback-modal__overlay {
+.feedback-modal-iframe {
+    display: none;
     position: fixed;
     inset: 0;
-    z-index: 1001;
-    background: rgba(0, 0, 0, 0.5);
-    display: flex;
-    align-items: center;
-    justify-content: center;
-    opacity: 0;
-    pointer-events: none;
-    transition: opacity 0.2s;
+    width: 100%;
+    height: 100%;
+    border: none;
+    z-index: 9999;
 }
 
-.feedback-modal__overlay--visible {
-    opacity: 1;
-    pointer-events: auto;
+.feedback-modal-iframe--visible {
+    display: block;
 }
 
 .feedback-modal {
@@ -551,6 +547,55 @@
     padding: 40px 0;
     color: var(--accent);
     font-size: 14px;
+}
+
+.feedback-modal__messages {
+    display: flex;
+    flex-direction: column;
+    gap: 10px;
+    margin-bottom: 16px;
+}
+
+.feedback-modal__message {
+    padding: 10px 14px;
+    border-radius: 8px;
+    max-width: 85%;
+}
+
+.feedback-modal__message--user {
+    background: var(--light-grey);
+    align-self: flex-start;
+}
+
+.feedback-modal__message--admin {
+    background: #e0f2f1;
+    align-self: flex-end;
+}
+
+.feedback-modal__message-meta {
+    display: flex;
+    gap: 8px;
+    align-items: center;
+    margin-bottom: 4px;
+}
+
+.feedback-modal__message-author {
+    font-size: 11px;
+    font-weight: 600;
+    color: var(--teal-green);
+}
+
+.feedback-modal__message-date {
+    font-size: 11px;
+    color: var(--accent);
+}
+
+.feedback-modal__message-text {
+    font-size: 13px;
+    color: var(--dark-primary);
+    line-height: 1.5;
+    white-space: pre-wrap;
+    word-break: break-word;
 }
 
 @media (max-width: 768px) {
