+++ src/widgets/feedback-modal/FeedbackModal.scss
@@ -1,19 +1,15 @@
-.feedback-modal__overlay {
+.feedback-modal-iframe {
+    display: none;
     position: fixed;
     inset: 0;
-    z-index: 1001;
-    background: rgba(0, 0, 0, 0.5);
-    display: flex;
-    align-items: center;
-    justify-content: center;
-    opacity: 0;
-    pointer-events: none;
-    transition: opacity 0.2s;
+    width: 100%;
+    height: 100%;
+    border: none;
+    z-index: 9999;
 }
 
-.feedback-modal__overlay--visible {
-    opacity: 1;
-    pointer-events: auto;
+.feedback-modal-iframe--visible {
+    display: block;
 }
 
 .feedback-modal {
